Move SubscriptionService to top level API package
