+++ pkg/northbound/subscription/service.go
@@ -9,7 +9,7 @@ import (
 	substore "github.com/onosproject/onos-e2t/pkg/store/subscription"
 	"io"
 
-	subapi "github.com/onosproject/onos-e2t/api/ricapi/e2/subscription/v1beta1"
+	subapi "github.com/onosproject/onos-e2t/api/subscription/v1beta1"
 	"github.com/onosproject/onos-lib-go/pkg/logging"
 	"github.com/onosproject/onos-lib-go/pkg/northbound"
 	"google.golang.org/grpc"
